third_party/blink/web_tests/http/tests/devtools/elements/styles/original-content-provider-expected.txt

Verifies that CSSStyleSheetHeader.originalContentProvider() indeed returns original content.


Running: testSetStyle
original content loaded for header: set-style.css
== Original ==
div { color: red; }
== Current ==
div {EDITED: EDITED}

Running: testSetSelector
original content loaded for header: set-selector.css
== Original ==
div {}
== Current ==
EDITED {}

Running: testSetMedia
original content loaded for header: set-media.css
== Original ==
@media (all) { }
== Current ==
@media EDITED { }

Running: testSetKeyframeKey
original content loaded for header: set-keyframe-key.css
== Original ==
@keyframes animation { 100% { color: red; } }
== Current ==
@keyframes animation { from { color: red; } }

Running: testAddRule
original content loaded for header: add-rule.css
== Original ==
div {}
== Current ==
EDITED {}
div {}

Running: testSetStyleSheetText
original content loaded for header: set-text.css
== Original ==
div {}
== Current ==
EDITED {}
